refactor: canonicalize PresetDatabase interface, fix App db stability, add webmidi mock
Agent-Logs-Url: https://github.com/fpbrault/cz101-presets-library/sessions/acc38661-2699-44e7-8ba8-d1f53ca0866e
