+++ setupBrowserTests.ts
@@ -4,6 +4,20 @@ import "@testing-library/jest-dom/vitest";
 import { beforeEach, vi } from "vitest";
 import "./src/App.css";
 
+// Mock the webmidi module
+vi.mock("webmidi", () => ({
+	WebMidi: {
+		enabled: false,
+		inputs: [],
+		outputs: [],
+		enable: vi.fn().mockResolvedValue(undefined),
+		addListener: vi.fn(),
+		removeListener: vi.fn(),
+		getInputByName: vi.fn().mockReturnValue(undefined),
+		getOutputByName: vi.fn().mockReturnValue(undefined),
+	},
+}));
+
 // Mock the MIDIAccess object
 const mockMIDIAccess = {
 	inputs: new Map(),
