+++ src/lib/db/PresetDatabase.ts
@@ -1,7 +1,10 @@
-import type { Preset } from "@/lib/presets/presetManager";
+import type { Preset } from "@/lib/presets/types";
+
+export type { Preset };
 
 export interface PresetDatabase {
 	getPresets(): Promise<Preset[]>;
+	getPresetById(id: string): Promise<Preset | null>;
 	addPreset(preset: Preset): Promise<Preset>;
 	updatePreset(preset: Preset): Promise<void>;
 	deletePreset(id: string): Promise<void>;
